Do not fail task creation immediately on status polling failure
When task is being created, a django_rq job is created and scheduled to the worker.
While the worker is processing the job, front-end polls back-end for the job status.
If any poll request fails (even if due to network error),
front-end assumes task creation has failed and deletes the job (it happens quite a lot for big tasks).

Fix it by concluding that task creation has failed not on the first error response,
but on the serie of 7 subsequent error responses.